Refactor logic module and gates for better organization and modularity
