+++ src/logic/mod.rs
@@ -1,15 +1,16 @@
 use bevy::prelude::*;
-use derive_more::From;
 
 #[allow(unused_imports)]
 pub mod prelude {
     pub use super::components::*;
-    pub use super::gates::{ AndGate };
+    pub use super::gates::{ NotGate, AndGate, OrGate };
+    pub use super::signal::Signal;
 }
 
+pub mod signal;
 pub mod gates;
 pub mod components {
-    use super::*;
+    use super::{ *, signal::Signal };
 
     /// An output node.
     #[derive(Component, Clone, Copy, Debug)]
@@ -31,51 +32,20 @@ pub mod components {
     }
 }
 
-/// State storage for logic simulation.
-#[derive(Clone, Copy, Debug, PartialEq, PartialOrd, From)]
-pub enum Signal {
-    Analog(f32),
-    Digital(bool),
-    Undefined,
-}
+fn debug_logic_components(
+    mut gizmos: Gizmos,
+    query_gates: Query<(&dyn gates::LogicGate, &Children)>,
+    query_sources: Query<&components::Source>,
+    query_sinks: Query<&components::Sink>,
+    query_wires: Query<&components::Wire>
+) {
+    for (gate, children) in query_gates.iter() {
+        let sources = children.iter().filter_map(|entity| query_sources.get(*entity).ok());
+        let sinks = children.iter().filter_map(|entity| query_sinks.get(*entity).ok());
+        let wires = children.iter().filter_map(|entity| query_wires.get(*entity).ok());
 
-impl Signal {
-    pub const OFF: Signal = Signal::Digital(false);
-    pub const ON: Signal = Signal::Digital(true);
+        // todo: spatial bundles on gates
 
-    /// Returns true if the signal is true or greater or equal to 1.0.
-    pub fn is_true(&self) -> bool {
-        match self {
-            Signal::Digital(true) => true,
-            Signal::Analog(value) => *value >= 1.0,
-            _ => false,
-        }
-    }
-}
-
-impl std::ops::Add for Signal {
-    type Output = Self;
-
-    fn add(self, rhs: Self) -> Self::Output {
-        match (self, rhs) {
-            (Signal::Undefined, _) | (_, Signal::Undefined) => Signal::Undefined,
-            (Signal::Analog(lhs), Signal::Analog(rhs)) => Signal::Analog(lhs + rhs),
-            (Signal::Analog(a), Signal::Digital(d)) | (Signal::Digital(d), Signal::Analog(a)) =>
-                Signal::Analog(a + (if d { 1.0 } else { 0.0 })),
-
-            (Signal::Digital(lhs), Signal::Digital(rhs)) => Signal::Digital(lhs || rhs),
-        }
-    }
-}
-
-impl std::ops::Add<f32> for Signal {
-    type Output = Self;
-    fn add(self, rhs: f32) -> Self::Output {
-        match self {
-            Signal::Analog(value) => Signal::Analog(value + rhs),
-            Signal::Digital(true) => Signal::Analog(1.0 + rhs),
-            Signal::Digital(false) => Signal::Analog(rhs),
-            Signal::Undefined => Signal::Undefined,
-        }
+        // gizmos.circle(position, normal, radius, color)
     }
 }
